requirements.txt
openai
langchain
langchain-community
chromadb
tiktoken
ipywidgets
pypdf
pdfplumber
python-dotenv
fastapi
uvicorn
python-multipart
langchain
openai
pypdf
